+++ Note/CHANGELOG.md
@@ -6,6 +6,12 @@ Tổng hợp các cập nhật, nâng cấp và sửa lỗi theo từng ngày ph
 
 > Theo chuẩn Semantic Versioning (SemVer)
 ---
+## [1.3.1] - 2025-05-01
+- cấu trúc lại folder
+- refactor code
+
+---
+
 ## [1.3.0] - 2025-05-01
 
 ### ✨ Tính năng mới
