+++ frontend/public/select.html
@@ -3,7 +3,8 @@
   <head>
     <meta charset="UTF-8" />
     <title>Chọn Manga Folder</title>
-    <link rel="stylesheet" href="/src/styles/select.css" />
+    <link rel="stylesheet" href="/src/styles/base.css" />
+    <link rel="stylesheet" href="/src/styles/pages/select.css" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
   </head>
   <body>
@@ -15,6 +16,6 @@
       <!-- JS sẽ render danh sách folder vào đây -->
     </main>
 
-    <script type="module" src="/src/select.js"></script>
+    <script type="module" src="/src/pages/select.js"></script>
   </body>
 </html>
