+++ frontend/src/styles/base.css
@@ -1,71 +1,139 @@
-/* 📁 frontend/src/styles/base.css */
-body {
-  font-family: "Segoe UI", sans-serif;
-  background-color: #f2f2f2;
+/* =============================
+   🔧 RESET & GLOBAL
+============================= */
+
+* {
+  box-sizing: border-box;
   margin: 0;
   padding: 0;
+}
+
+html, body {
+  font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, Helvetica, Arial, sans-serif;
+  background-color: #fafafa;
   color: #333;
-}
-
-#wrapper {
-  padding: 20px;
-}
-
-.search {
-  margin: 40px 0 20px;
-  width: 100%;
-  display: flex;
-  justify-content: center;
-}
-
-.search input {
-  padding: 10px;
+  line-height: 1.5;
   font-size: 16px;
-  width: 300px;
-  max-width: 90vw;
-  box-sizing: border-box;
+  scroll-behavior: smooth;
 }
 
-.btn-fab {
-  position: fixed;
-  width: 48px;
-  height: 48px;
-  border-radius: 50%;
-  font-size: 24px;
-  font-weight: bold;
-  background-color: rgba(0, 0, 0, 0.7);
-  color: white;
-  border: none;
-  z-index: 999;
+/* Ẩn phần bị JS toggle */
+.hidden {
+  display: none !important;
+}
+
+/* =============================
+   🎛 BUTTON CƠ BẢN / TOÀN CỤC
+============================= */
+
+button {
+  font-family: inherit;
+  font-size: inherit;
   cursor: pointer;
-  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.3);
+  transition: background 0.2s ease;
+}
+
+button:disabled {
+  opacity: 0.5;
+  cursor: not-allowed;
+}
+
+/* =============================
+   📦 PRELOAD / LOADING STATE
+============================= */
+
+img.loading {
+  opacity: 0.3;
+  filter: blur(2px);
+  transition: opacity 0.3s ease, filter 0.3s ease;
+}
+
+/* =============================
+   🌀 OVERLAY LOADING
+============================= */
+
+#loading-overlay {
+  position: fixed;
+  inset: 0;
+  z-index: 99999;
+  background: rgba(0, 0, 0, 0.6);
+  color: white;
+  font-size: 18px;
   display: flex;
   align-items: center;
   justify-content: center;
-  transition: opacity 0.3s ease;
+  font-weight: bold;
+  backdrop-filter: blur(6px);
 }
 
-.btn-fab:hover {
-  background-color: rgba(0, 0, 0, 0.9);
+#loading-overlay.hidden {
+  display: none !important;
 }
 
-.back-btn {
-  top: 20px;
-  left: 20px;
+/* =============================
+   🔍 TÌM KIẾM NỔI FLOATING
+============================= */
+
+.floating-search {
+  position: fixed;
+  top: 56px;
+  left: 0;
+  width: 100%;
+  background: white;
+  z-index: 1000;
+  padding: 8px 12px;
+  box-shadow: 0 2px 10px rgba(0,0,0,0.15);
+  display: none;
 }
 
-.mode-btn {
-  top: 20px;
-  right: 80px;
+.floating-search.active {
+  display: block;
 }
 
-.dark-btn {
-  top: 20px;
-  right: 20px;
+.floating-search input {
+  width: 100%;
+  font-size: 16px;
+  padding: 10px;
+  border-radius: 8px;
+  border: 1px solid #ccc;
+  outline: none;
 }
 
+#search-dropdown {
+  margin-top: 6px;
+  background: #fff;
+  border-radius: 8px;
+  overflow: hidden;
+  box-shadow: 0 2px 12px rgba(0,0,0,0.2);
+  max-height: 300px;
+  overflow-y: auto;
+  display: none;
+}
 
-/* 📚 Hiệu ứng khi hover vào logo */
-#logo-link:hover {
-  opacity: 0.7;
+#search-dropdown:not(.hidden) {
+  display: block;
+}
+
+.search-item {
+  display: flex;
+  gap: 12px;
+  align-items: center;
+  padding: 10px;
+  border-bottom: 1px solid #eee;
+}
+
+.search-item:hover {
+  background: #f2f2f2;
+}
+
+.search-thumb {
+  width: 60px;
+  height: 60px;
+  object-fit: cover;
+  border-radius: 8px;
+}
+
+.search-title {
+  font-weight: bold;
+  font-size: 16px;
 }
