--- frontend/src/styles/dark.css
@@ -1,20 +1,0 @@
-/* 📁 frontend/src/styles/dark.css */
-body.dark-mode {
-    background-color: #121212;
-    color: #eee;
-  }
-  
-  .dark-mode .card {
-    background: #1e1e1e;
-    box-shadow: 0 4px 12px rgba(255, 255, 255, 0.05);
-  }
-  
-  .dark-mode .reader {
-    background-color: #000;
-  }
-  
-  .dark-mode .page-indicator {
-    background-color: rgba(255, 255, 255, 0.2);
-    color: #fff;
-  }
-
