+++ frontend/src/styles/dark/home-dark.css
@@ -1,0 +1,63 @@
+/* =============================
+   🌙 DARK MODE – CHO TRANG HOME (index.html)
+============================= */
+
+/* Header tối */
+.dark-mode #site-header {
+    background: rgba(20, 20, 20, 0.9);
+    color: white;
+  }
+  
+  /* Sidebar tối */
+  .dark-mode #sidebar-menu {
+    background: #1f1f1f;
+  }
+  
+  /* Card folder */
+  .dark-mode .card {
+    background: #2a2a2a;
+    color: #f1f1f1;
+  }
+  
+  .dark-mode .card div {
+    color: #eee;
+  }
+  
+  .dark-mode .card:hover {
+    box-shadow: 0 8px 16px rgba(255, 255, 255, 0.1);
+  }
+  
+  /* Nút phân trang */
+  .dark-mode .reader-controls button {
+    background: #333;
+    color: white;
+  }
+  
+  .dark-mode .reader-controls button:hover {
+    background: #444;
+  }
+  
+  /* Nút sidebar */
+  .dark-mode #sidebar-menu button:hover {
+    background: #333;
+  }
+  
+  /* Footer */
+  .dark-mode #main-footer {
+    background: rgba(0, 0, 0, 0.8);
+  }
+  
+  /* Search dropdown */
+  .dark-mode #search-dropdown {
+    background: #222;
+    color: white;
+  }
+  
+  .dark-mode .search-item:hover {
+    background: #333;
+  }
+  
+  .dark-mode .search-title {
+    color: #eee;
+  }
+
