--- frontend/src/styles/folder.css
@@ -1,71 +1,0 @@
-/* 📁 frontend/src/styles/folder.css */
-
-/* Lưới grid folder */
-.grid {
-  display: grid;
-  grid-template-columns: repeat(auto-fill, minmax(160px, 1fr)); /* ✅ Tự co giãn card */
-  gap: 16px;
-  padding: 16px;
-}
-
-/* Card folder */
-.card {
-  background: var(--card-bg, #fff); /* ✅ Hỗ trợ dark mode */
-  border-radius: 12px;
-  box-shadow: 0 4px 8px rgba(0,0,0,0.08);
-  overflow: hidden;
-  transition: transform 0.2s, box-shadow 0.2s;
-  text-align: center;
-  font-size: 14px;
-  cursor: pointer;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-/* Hover nhẹ */
-.card:hover {
-  transform: translateY(-6px); /* ✅ Hover nổi lên mượt */
-  box-shadow: 0 8px 16px rgba(0,0,0,0.15);
-}
-
-/* Ảnh thumbnail trong card */
-.card img {
-  width: 100%;
-  height: 220px;
-  object-fit: cover; /* ✅ Ảnh luôn đầy nhưng không méo */
-  border-bottom: 1px solid #ddd;
-}
-
-/* Tên folder dưới ảnh */
-.card div {
-  padding: 10px;
-  font-weight: bold;
-  font-size: 16px;
-  color: var(--text-color, #333); /* ✅ Tự đổi khi dark mode */
-  text-align: center;
-}
-
-/* Thanh điều khiển phân trang */
-.reader-controls {
-  display: flex;
-  justify-content: center;
-  gap: 20px;
-  margin: 20px;
-}
-
-/* Nút phân trang */
-.reader-controls button {
-  padding: 10px 16px;
-  font-size: 16px;
-  border: none;
-  border-radius: 8px;
-  background: var(--btn-bg, #eee);
-  cursor: pointer;
-  transition: background 0.2s;
-}
-
-/* Hover nút phân trang */
-.reader-controls button:hover {
-  background: var(--btn-hover-bg, #ccc);
-}
