--- frontend/src/styles/layout.css
@@ -1,278 +1,0 @@
-/* ================================
-   📌 HEADER – YouTube Style + Blur
-================================== */
-#site-header {
-  position: fixed;
-  top: 0;
-  left: 0;
-  right: 0;
-  z-index: 1000;
-  background: rgba(255, 255, 255, 0.75);
-  backdrop-filter: blur(12px);
-  -webkit-backdrop-filter: blur(12px);
-  color: black;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 10px 16px;
-  box-shadow: 0 1px 6px rgba(0, 0, 0, 0.1);
-}
-
-.header-left h1 {
-  font-size: 20px;
-  font-weight: 600;
-  margin: 0;
-}
-
-.header-icons {
-  display: flex;
-  align-items: center;
-  gap: 10px;
-}
-
-.header-icons button {
-  background: none;
-  border: none;
-  font-size: 18px;
-  cursor: pointer;
-  color: inherit;
-  padding: 6px 8px;
-  border-radius: 6px;
-}
-
-.header-icons button:hover {
-  background-color: rgba(0, 0, 0, 0.05);
-}
-
-/* ================================
-   🔍 2. TÌM KIẾM – Thanh trượt xuống kiểu YouTube
-================================== */
-#floating-search {
-  position: fixed;
-  top: 0;
-  left: 0;
-  right: 0;
-  background: white;
-  padding: 12px 16px;
-  display: flex;
-  gap: 10px;
-  align-items: center;
-  transform: translateY(-100%);
-  transition: transform 0.3s ease;
-  z-index: 1001;
-  box-shadow: 0 4px 10px rgba(0, 0, 0, 0.15);
-}
-
-#floating-search.active {
-  transform: translateY(0);
-}
-
-#floating-search input {
-  flex: 1;
-  padding: 10px;
-  font-size: 16px;
-  border: 1px solid #ccc;
-  border-radius: 6px;
-  outline: none;
-}
-#search-dropdown {
-  position: absolute;
-  top: 86px; /* dưới thanh search nổi */
-  left: 16px;
-  width: calc(100% - 32px);
-  background: white;
-  box-shadow: 0 4px 12px rgba(0,0,0,0.15);
-  border-radius: 8px;
-  max-height: 300px;
-  overflow-y: auto;
-  z-index: 1001;
-}
-
-.search-item {
-  display: flex;
-  gap: 10px;
-  padding: 8px 12px;
-  cursor: pointer;
-  align-items: center;
-}
-
-.search-item:hover {
-  background: #f0f0f0;
-}
-
-.search-thumb {
-  width: 40px;
-  height: 40px;
-  border-radius: 4px;
-  object-fit: cover;
-}
-
-.search-title {
-  font-weight: bold;
-  font-size: 14px;
-  color: #333;
-}
-
-#search-loader {
-  padding: 10px;
-  text-align: center;
-  font-size: 14px;
-  color: #777;
-}
-
-/* ================================
-   📦 3. WRAPPER – Flex layout + né header
-================================== */
-html,
-body {
-  height: 100%;
-  margin: 0;
-  padding: 0;
-}
-
-#wrapper {
-  min-height: 100vh;
-  display: flex;
-  flex-direction: column;
-  padding-top: 80px;
-  /* né header */
-}
-
-#app {
-  flex: 1;
-  /* nội dung chính sẽ đẩy footer xuống cuối */
-}
-
-.reader-mode #wrapper {
-  padding: 0 !important;
-}
-
-/* ================================
-   🔻 4. FOOTER – Không còn fixed
-================================== */
-#site-footer,
-#main-footer {
-  background: rgba(0, 0, 0, 0.6);
-  backdrop-filter: blur(10px);
-  -webkit-backdrop-filter: blur(10px);
-  color: white;
-  text-align: center;
-  font-size: 13px;
-  padding: 12px 20px;
-  margin-top: 40px;
-}
-
-/* ================================
-   🕶️ 5. UTILITIES – Các lớp phụ
-================================== */
-.hidden {
-  display: none !important;
-}
-
-/* ================================
-   📱 6. RESPONSIVE – Mobile tweaks
-================================== */
-@media (max-width: 600px) {
-  #site-header {
-    flex-direction: column;
-    align-items: flex-start;
-  }
-
-  .header-icons {
-    justify-content: flex-end;
-    width: 100%;
-    margin-top: 10px;
-  }
-}
-
-
-/* ================================
-   📚 FOOTER riêng cho chế độ reader
-================================== */
-#reader-footer {
-  position: fixed;
-  bottom: 0;
-  left: 0;
-  right: 0;
-  z-index: 999;
-  background: rgba(0, 0, 0, 0.7);
-  backdrop-filter: blur(8px);
-  color: white;
-  padding: 10px 16px;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.reader-footer-inner {
-  width: 100%;
-  max-width: 800px;
-  display: flex;
-  justify-content: space-between; /* 📌 Trái - giữa - phải */
-  align-items: center;
-  margin: 0 auto;
-}
-#reader-footer button {
-  background: #ffffff22;
-  color: white;
-  border: none;
-  padding: 6px 14px;
-  font-size: 14px;
-  border-radius: 6px;
-  cursor: pointer;
-}
-
-#reader-footer button:hover {
-  background: #ffffff33;
-}
-
-#page-info {
-  flex-grow: 1;
-  text-align: center; /* ✅ Căn giữa số trang */
-  font-weight: bold;
-}
-
-/* Ẩn khi cần */
-.hidden {
-  display: none !important;
-}
-
-body.reader-mode #main-footer {
-  display: none !important;
-  /* 🔥 Ẩn triệt để khi đọc truyện */
-}
-
-
-/* ================================
-   🔄 Loading Overlay
-================================== */
-#loading-overlay {
-  position: fixed;
-  top: 0;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  background: rgba(255, 255, 255, 0.7);
-  backdrop-filter: blur(4px);
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  z-index: 2000;
-}
-
-#loading-overlay.hidden {
-  display: none;
-}
-
-.spinner {
-  width: 50px;
-  height: 50px;
-  border: 5px solid #ccc;
-  border-top: 5px solid #333;
-  border-radius: 50%;
-  animation: spin 1s linear infinite;
-}
-
-@keyframes spin {
-  to { transform: rotate(360deg); }
-}
