+++ frontend/src/styles/pages/home.css
@@ -1,0 +1,188 @@
+/* =============================
+   🏠 HOME PAGE – HEADER, FOOTER, SIDEBAR
+============================= */
+
+/* Header cố định trên đầu */
+#site-header {
+    position: fixed;
+    top: 0;
+    left: 0;
+    right: 0;
+    z-index: 1000;
+    background: rgba(255, 255, 255, 0.75);
+    backdrop-filter: blur(12px);
+    color: black;
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    padding: 10px 16px;
+    box-shadow: 0 1px 6px rgba(0, 0, 0, 0.1);
+  }
+  
+  /* Icon bên phải header */
+  .header-icons {
+    display: flex;
+    align-items: center;
+    gap: 10px;
+  }
+  
+  .header-icons button {
+    background: none;
+    border: none;
+    font-size: 18px;
+    cursor: pointer;
+    color: inherit;
+    padding: 6px 8px;
+    border-radius: 6px;
+  }
+  
+  .header-icons button:hover {
+    background-color: rgba(0, 0, 0, 0.05);
+  }
+  
+  /* Wrapper toàn trang */
+  #wrapper {
+    min-height: 100vh;
+    display: flex;
+    flex-direction: column;
+    padding-top: 80px; /* né header */
+  }
+  
+  #app {
+    flex: 1;
+  }
+  
+  /* Footer chính */
+  #main-footer {
+    background: rgba(0, 0, 0, 0.6);
+    backdrop-filter: blur(10px);
+    color: white;
+    text-align: center;
+    font-size: 13px;
+    padding: 12px 20px;
+    margin-top: 40px;
+  }
+  
+  /* =============================
+     📁 GRID FOLDER + CARD
+  ============================= */
+  
+  .grid {
+    display: grid;
+    grid-template-columns: repeat(auto-fill, minmax(160px, 1fr));
+    gap: 16px;
+    padding: 16px;
+  }
+  
+  .card {
+    background: var(--card-bg, #fff);
+    border-radius: 12px;
+    box-shadow: 0 4px 8px rgba(0,0,0,0.08);
+    overflow: hidden;
+    transition: transform 0.2s, box-shadow 0.2s;
+    text-align: center;
+    font-size: 14px;
+    cursor: pointer;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+  }
+  
+  .card:hover {
+    transform: translateY(-6px);
+    box-shadow: 0 8px 16px rgba(0,0,0,0.15);
+  }
+  
+  .card img {
+    width: 100%;
+    height: 220px;
+    object-fit: cover;
+    border-bottom: 1px solid #ddd;
+  }
+  
+  .card div {
+    padding: 10px;
+    font-weight: bold;
+    font-size: 16px;
+    color: var(--text-color, #333);
+  }
+  
+  /* =============================
+     📄 PAGINATION
+  ============================= */
+  .reader-controls {
+    display: flex;
+    justify-content: center;
+    gap: 20px;
+    margin: 20px;
+  }
+  
+  .reader-controls button {
+    padding: 10px 16px;
+    font-size: 16px;
+    border: none;
+    border-radius: 8px;
+    background: var(--btn-bg, #eee);
+    cursor: pointer;
+    transition: background 0.2s;
+  }
+  
+  .reader-controls button:hover {
+    background: var(--btn-hover-bg, #ccc);
+  }
+  
+  /* =============================
+     📚 SIDEBAR MENU
+  ============================= */
+  
+  #sidebar-menu {
+    position: fixed;
+    top: 60px;
+    left: 0;
+    width: 0;
+    height: calc(100% - 60px);
+    background: #2e2e2e;
+    overflow-y: auto;
+    transition: 0.3s;
+    z-index: 999;
+    padding-top: 20px;
+    color: white;
+    box-shadow: 2px 0 5px rgba(0,0,0,0.3);
+  }
+  
+  #sidebar-menu.active {
+    width: 250px;
+  }
+  
+  #sidebar-menu button {
+    display: block;
+    background: none;
+    border: none;
+    color: white;
+    text-align: left;
+    padding: 12px 20px;
+    font-size: 16px;
+    width: 100%;
+    cursor: pointer;
+  }
+  
+  #sidebar-menu button:hover {
+    background: #444;
+  }
+  
+  /* =============================
+     🔍 RESPONSIVE
+  ============================= */
+  @media (max-width: 600px) {
+    #site-header {
+      flex-direction: column;
+      align-items: flex-start;
+    }
+  
+    .header-icons {
+      justify-content: flex-end;
+      width: 100%;
+      margin-top: 10px;
+    }
+  }
+
