+++ frontend/src/styles/pages/reader.css
@@ -1,0 +1,180 @@
+/* =============================
+   📖 READER PAGE – CHẾ ĐỘ ĐỌC TRUYỆN
+============================= */
+
+/* Tắt padding wrapper khi vào reader */
+/* body.reader-mode #wrapper {
+  padding: 0 !important;
+} */
+
+
+/* =============================
+   📌 FIXED HEADER CHO reader.html
+============================= */
+#site-header {
+  position: fixed;
+  top: 0;
+  left: 0;
+  right: 0;
+  z-index: 999;
+  background: rgba(20, 20, 20, 0.85);
+  backdrop-filter: blur(12px);
+  color: white;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  padding: 10px 16px;
+  box-shadow: 0 1px 6px rgba(0, 0, 0, 0.2);
+  transition: transform 0.3s ease;
+}
+
+#site-header.hidden {
+  transform: translateY(-100%);
+}
+
+/*  */
+/* Khung reader chính */
+.reader {
+  position: relative;
+  height: 100vh;
+  overflow: hidden;
+  background-color: #000;
+}
+
+/* Ảnh 1 trang (swipe mode) */
+.reader:not(.scroll-mode) img {
+  width: 100%;
+  height: 100%;
+  object-fit: contain;
+  display: none;
+  margin: 0 auto;
+  position: absolute;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+}
+
+/* Scroll mode */
+.reader.scroll-mode {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  overflow-y: auto;
+  height: auto;
+  padding: 0;
+  background-color: #000;
+}
+
+/* =============================
+   📷 ẢNH SCROLL MODE
+============================= */
+.scroll-img {
+  width: 100%;
+  height: auto;
+  object-fit: contain;
+  margin-bottom: 20px;
+  display: block;
+  -webkit-overflow-scrolling: touch;
+}
+
+@media (max-width: 768px) {
+  .scroll-img {
+    width: 100vw;
+    max-width: 100vw;
+  }
+}
+
+/* =============================
+   🧭 TRẠNG THÁI TRANG (hiện ở giữa màn hình)
+============================= */
+.page-indicator {
+  position: absolute;
+  bottom: 10px;
+  left: 50%;
+  transform: translateX(-50%);
+  color: white;
+  font-size: 16px;
+  background-color: rgba(0, 0, 0, 0.5);
+  padding: 4px 10px;
+  border-radius: 12px;
+}
+
+/* =============================
+   💫 ẢNH LOADING
+============================= */
+.scroll-img.loading,
+.reader img.loading {
+  opacity: 0.3;
+  filter: blur(2px);
+  transition: opacity 0.5s ease, filter 0.5s ease;
+}
+
+/* =============================
+   📦 FOOTER ĐỌC TRUYỆN
+============================= */
+#reader-footer {
+  position: fixed;
+  bottom: 0;
+  left: 0;
+  right: 0;
+  z-index: 999;
+  background: rgba(0, 0, 0, 0.7);
+  backdrop-filter: blur(8px);
+  color: white;
+  padding: 10px 16px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+.reader-footer-inner {
+  width: 100%;
+  max-width: 800px;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  margin: 0 auto;
+}
+
+#reader-footer button {
+  background: #ffffff22;
+  color: white;
+  border: none;
+  padding: 6px 14px;
+  font-size: 14px;
+  border-radius: 6px;
+  cursor: pointer;
+}
+
+#reader-footer button:hover {
+  background: #ffffff33;
+}
+
+#page-info {
+  flex: 1;
+  text-align: center;
+  font-weight: bold;
+  font-size: 15px;
+  padding: 4px 0;
+}
+
+
+
+/* Count page trong scroll */
+#image-count-info {
+  position: fixed;
+  bottom: 60px;
+  left: 50%;
+  transform: translateX(-50%);
+  background: rgba(0,0,0,0.6);
+  color: white;
+  padding: 6px 12px;
+  border-radius: 8px;
+  z-index: 999;
+  font-weight: bold;
+}
+
+/* Ẩn image count khi không ở scroll mode */
+.reader:not(.scroll-mode) #image-count-info {
+  display: none;
+}
