--- frontend/src/styles/reader.css
@@ -1,79 +1,0 @@
-/* 📁 frontend/src/styles/reader.css */
-body.reader-mode #wrapper {
-  padding: 0 !important;
-}
-
-.reader {
-  position: relative;
-  height: 100vh;
-  overflow: hidden;
-  background-color: #000;
-}
-
-.reader:not(.scroll-mode) img {
-  width: 100%;
-  height: 100%;
-  object-fit: contain;
-  display: none;
-  margin: 0 auto;
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-}
-
-.reader.scroll-mode {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  overflow-y: auto;
-  height: auto;
-  padding: 0;
-  background-color: #000;
-}
-
-.scroll-img {
-  /* max-width: 100%;   => ảnh gốc */
-  width: 100%;  /*ảnh bé cho max with*/
-  height: auto;
-  object-fit: contain;
-  margin-bottom: 20px;
-  display: block;
-  -webkit-overflow-scrolling: touch;
-
-}
-
-@media (max-width: 768px) {
-  .scroll-img {
-    width: 100vw;
-    max-width: 100vw;
-  }
-}
-
-.page-indicator {
-  position: absolute;
-  bottom: 10px;
-  left: 50%;
-  transform: translateX(-50%);
-  color: white;
-  font-size: 16px;
-  background-color: rgba(0, 0, 0, 0.5);
-  padding: 4px 10px;
-  border-radius: 12px;
-}
-
-.scroll-img.loading {
-  opacity: 0.3;
-  filter: blur(2px);
-  transition: opacity 0.5s ease, filter 0.5s ease;
-}
-
-.reader img.loading {
-  opacity: 0.3;
-  filter: blur(2px);
-  transition: opacity 0.5s ease, filter 0.5s ease;
-}
-
-
-
-/*////////////////////////// */
