--- frontend/src/styles/sidebar.css
@@ -1,41 +1,0 @@
-/* 📁 frontend/src/styles/sidebar.css */
-
-/* Sidebar menu */
-#sidebar-menu {
-    position: fixed;
-    top: 60px; /* 💬 Dưới header */
-    left: 0;
-    width: 0; /* 🔥 Ban đầu ẩn */
-    height: calc(100% - 60px);
-    background: #2e2e2e;
-    overflow-x: hidden;
-    overflow-y: auto;
-    transition: 0.3s;
-    z-index: 999;
-    padding-top: 20px;
-    color: white;
-    box-shadow: 2px 0 5px rgba(0,0,0,0.3);
-  }
-  
-  /* Khi active sidebar */
-  #sidebar-menu.active {
-    width: 250px; /* 🔥 Khi mở */
-  }
-  
-  /* Các nút trong sidebar */
-  #sidebar-menu button {
-    display: block;
-    background: none;
-    border: none;
-    color: white;
-    text-align: left;
-    padding: 12px 20px;
-    font-size: 16px;
-    width: 100%;
-    cursor: pointer;
-  }
-  
-  #sidebar-menu button:hover {
-    background: #444;
-  }
-
