+++ package.json
@@ -18,6 +18,7 @@
     "better-sqlite3": "^11.9.1",
     "cors": "^2.8.5",
     "express": "^5.1.0",
+    "hammerjs": "^2.0.8",
     "serve-favicon": "^2.5.0",
     "string-natural-compare": "^3.0.1"
   }
